Moved test util functions back to _test files
- Had to do this because deadcode tool doesn't check
  *_test files for use of these functions
